wip: passing some computed tests
